Fix AutoIncrement Problem on PostGres Provider

+++ PTZ.HomeManagement.MyFinance.Services/Interfaces/IMyFinanceService.cs
@@ -9,20 +9,20 @@ namespace PTZ.HomeManagement.MyFinance
     public interface IMyFinanceService
     {
         BankAccount GetBankAccountDefault(string userId);
-        BankAccountMovement GetBankAccountMovementDefault(string userId, int bankAccountId);
+        BankAccountMovement GetBankAccountMovementDefault(string userId, long bankAccountId);
 
-        BankAccount GetBankAccount(string userId, int bankAccountId);
+        BankAccount GetBankAccount(string userId, long bankAccountId);
         List<BankAccount> GetBankAccounts(string userId);
         void SaveBankAccount(string userId, BankAccount bankAccount);
         void DeleteBankAccount(string userId, BankAccount bankAccount);
 
-        BankAccountMovement GetBankAccountMovement(string userId, int bankAccountId, int movementId);
-        List<BankAccountMovement> GetBankAccountMovements(string userId, int bankAccountId, DateTime startDate, DateTime endDate);
-        List<BankAccountMovement> GetBankAccountMovements(string userId, int bankAccountId, int qtdOfMovements = 1000, SortOrder dateSortOrder = SortOrder.Unspecified);
-        void SaveBankAccountMovement(string userId, int bankAccountId, BankAccountMovement bankAccountMovement);
-        List<BankAccountMovement> SaveBankAccountMovements(string userId, int bankAccountId, List<BankAccountMovement> list);
-        void DeleteBankAccountMovement(string userId, int bankAccountId, BankAccountMovement bankAccountMovement);
+        BankAccountMovement GetBankAccountMovement(string userId, long bankAccountId, int movementId);
+        List<BankAccountMovement> GetBankAccountMovements(string userId, long bankAccountId, DateTime startDate, DateTime endDate);
+        List<BankAccountMovement> GetBankAccountMovements(string userId, long bankAccountId, int qtdOfMovements = 1000, SortOrder dateSortOrder = SortOrder.Unspecified);
+        void SaveBankAccountMovement(string userId, long bankAccountId, BankAccountMovement bankAccountMovement);
+        List<BankAccountMovement> SaveBankAccountMovements(string userId, long bankAccountId, List<BankAccountMovement> list);
+        void DeleteBankAccountMovement(string userId, long bankAccountId, BankAccountMovement bankAccountMovement);
 
-        List<BankAccountMovement> ImportBankAccountMovement(string userId, int bankAccountId, BankAccountMovementImportType importType, IFormFile file);
+        List<BankAccountMovement> ImportBankAccountMovement(string userId, long bankAccountId, BankAccountMovementImportType importType, IFormFile file);
     }
 }
