+++ PTZ.HomeManagement/Models/MyFinanceViewModels/AccountMovementImportViewModel.cs
@@ -16,12 +16,12 @@ namespace PTZ.HomeManagement.Models.MyFinanceViewModels
             AvailableImportTypes = Enum.GetValues(typeof(BankAccountMovementImportType)).Cast<BankAccountMovementImportType>().Select(x => { return new SelectListItem() { Value = x.ToString(), Text = EnumExtensions.GetDescription(x) }; });
         }
 
-        public AccountMovementImportViewModel(int bankAccountId) : this()
+        public AccountMovementImportViewModel(long bankAccountId) : this()
         {
             BankAccountId = BankAccountId == 0 ? bankAccountId : BankAccountId;
         }
 
-        public int BankAccountId { get; set; }
+        public long BankAccountId { get; set; }
         public BankAccountMovementImportType ImportType { get; set; }
         public IEnumerable<SelectListItem> AvailableImportTypes { get; set; }
         public List<AccountMovementReviewListItemViewModel> Items { get; set; }
